+++ tests/js/unit/command_registry.test.js
@@ -43,6 +43,7 @@ function registryGlobals(extra = {}) {
     commandRegistryCategories: null,
     apiFetch: () => {},
     logClientError: () => {},
+    loadShellCatalogs: vi.fn(() => Promise.resolve()),
     setComposerValue: () => {},
     ...extra,
   }